requirements.txt
camelot-py == 0.11.0
python-docx == 1.1.2
tabula-py == 2.9.3